Add2

+++ src/main/java/com/watering/domain/VO/ScoreVO.java
@@ -22,9 +22,9 @@ public class ScoreVO {
     @ApiModelProperty("评估人姓名")
     private String name;
     //态度评分
-    private float attitude;
+    private Double attitude;
     //能力评分
-    private float ability;
+    private Double ability;
     //评价描述
     private String detail;
 }
